moved process email action, connexion file to out of mailers module and some other mailers module restructuring
