+++ opwen_email_server/integration/connexion.py
@@ -13,10 +13,12 @@ from opwen_email_server.integration.azure import get_auth
 from opwen_email_server.integration.azure import get_client_storage
 from opwen_email_server.integration.azure import get_email_storage
 from opwen_email_server.integration.azure import get_mailbox_storage
+from opwen_email_server.integration.azure import get_no_auth
 from opwen_email_server.integration.azure import get_pending_storage
 from opwen_email_server.integration.azure import get_raw_email_storage
 from opwen_email_server.integration.azure import get_user_storage
 from opwen_email_server.integration.celery import inbound_store
+from opwen_email_server.integration.celery import process_service_email
 from opwen_email_server.integration.celery import register_client
 from opwen_email_server.integration.celery import written_store
 from opwen_email_server.services.auth import AnyOfBasicAuth
@@ -29,6 +31,12 @@ email_receive = ReceiveInboundEmail(
     auth=get_auth(),
     raw_email_storage=get_raw_email_storage(),
     next_task=inbound_store.delay,
+)
+
+receive_service_email = ReceiveInboundEmail(
+    auth=get_no_auth(),
+    raw_email_storage=get_raw_email_storage(),
+    next_task=process_service_email,
 )
 
 client_write = UploadClientEmails(
